CDHSH-97 Moved dialog to seperate component and fixed issues for pull request
