+++ services/frontend/src/components/DetailsDialog/DetailsDialog.scss
@@ -3,8 +3,8 @@
 @import "styles/_mixins.scss";
 @import "styles/_media_queries.scss";
 
-.project-info-dialog {
-    .project-info-dialog-title {
+.details-dialog {
+    .details-dialog-title {
         padding-bottom: $default-spacing * 0.5;
 
         h6 {
