+++ services/frontend/src/components/DetailsDialog/index.js
@@ -1,0 +1,1 @@
+export {default} from "./DetailsDialog";
