--- services/frontend/src/components/PersonalDetailsDialog/PersonalDetailsDialog.scss
@@ -1,14 +1,0 @@
-@import "styles/_colors.scss";
-@import "styles/_dimens.scss";
-@import "styles/_mixins.scss";
-@import "styles/_media_queries.scss";
-
-.personal-details-dialog {
-    #personal-details-dialog-title {
-        padding-bottom: $default-spacing * 0.5;
-
-        h6 {
-            font-weight: 700;
-        }
-    }
-}
