--- services/frontend/src/components/PersonalDetailsDialog/index.js
@@ -1,1 +1,0 @@
-export {default} from "./PersonalDetailsDialog";
