--- services/frontend/src/components/ProjectInfoDialog/index.js
@@ -1,1 +1,0 @@
-export {default} from "./ProjectInfoDialog";
