+++ services/frontend/src/components/index.js
@@ -1,6 +1,8 @@
 export {default as AvatarIcon} from "./AvatarIcon";
 export {default as AuthBranding} from "./AuthBranding";
 export {default as AuthForm} from "./AuthForm";
+export {default as DetailsDialog} from "./DetailsDialog";
+export {default as DetailsForm} from "./DetailsForm";
 export {default as EditSkillsDialog} from "./EditSkillsDialog";
 export {default as EmailInput} from "./EmailInput";
 export {default as FilterHeader} from "./FilterHeader";
@@ -10,10 +12,7 @@ export {default as LoadingValidator} from "./LoadingValidator";
 export {default as Navbar} from "./Navbar";
 export {default as NavSidebar} from "./NavSidebar";
 export {default as ProfileCard} from "./ProfileCard";
-export {default as PersonalDetailsDialog} from "./PersonalDetailsDialog";
-export {default as DetailsForm} from "./DetailsForm";
 export {default as ProjectCard} from "./ProjectCard";
-export {default as ProjectInfoDialog} from "./ProjectInfoDialog";
 export {default as ScrollContainer} from "./ScrollContainer";
 export {default as SkillBadges} from "./SkillsBadges";
 export {default as TextBadge} from "./TextBadge";
